Split assert methods from InviteTest into base class
